Reland "Change MojoHandle to 64 bits"
This is a reland of commit 483d6301490012448fe38a3f8a5b374290e15986

The original CL was reverted for breaking IME on Chrome OS. This
has been resolved by giving IME a Mojo function table which still
uses 32-bit handle values to retain ABI compatibility for now.
This function table forwards to the regular one while extending
or truncating input and output handle values as needed.

Original change's description:
> Change MojoHandle to 64 bits
>
> We no longer have need to preserve Mojo Core ABI stability because
> consuming Mojo Core as a shared library is no longer supported. This
> changes MojoHandle to a 64-bit value in preparation for the
> introduction of ipcz -- which itself uses 64-bit handles -- making it
> trivial to reuse existing Mojo C++ handle types to wrap Mojo and ipcz
> handles alike.
>
> Note that this does not impact the wire format in any way, because
> handle values are only for the (process-local) API interface. In
> messages, handles are still encoded as 32-bit indices to some
> object metadata within the message.
>
> Bug: 1299283
> Binary-Size: This intentionally increases the size of a widely used data type from 32 bits to 64 bits, a change which is required to support other improvements. The 110 kB binary size increase is a direct result of that, and it does not seem to be avoidable.
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3493659
> Commit-Queue: Ken Rockot <rockot@google.com>
> Cr-Commit-Position: refs/heads/main@{#976468}

Bug: 1299283
Binary-Size: See rationale in the original commit message above.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501822
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#977729}

+++ content/browser/android/message_port_descriptor.cc
@@ -20,7 +20,7 @@ using base::android::ScopedJavaLocalRef;
 
 namespace {
 
-mojo::ScopedMessagePipeHandle WrapNativeHandle(jint native_handle) {
+mojo::ScopedMessagePipeHandle WrapNativeHandle(jlong native_handle) {
   MojoHandle raw_handle = static_cast<MojoHandle>(native_handle);
   DCHECK_NE(MOJO_HANDLE_INVALID, raw_handle);
   return mojo::ScopedMessagePipeHandle(mojo::MessagePipeHandle(raw_handle));
@@ -46,7 +46,7 @@ JNI_AppWebMessagePortDescriptor_CreatePair(JNIEnv* env) {
 }
 
 JNI_EXPORT jlong JNI_AppWebMessagePortDescriptor_Create(JNIEnv* env,
-                                                        jint native_handle,
+                                                        jlong native_handle,
                                                         jlong id_low,
                                                         jlong id_high,
                                                         jlong sequence_number) {
@@ -62,7 +62,7 @@ JNI_EXPORT jlong JNI_AppWebMessagePortDescriptor_Create(JNIEnv* env,
   return reinterpret_cast<jlong>(port);
 }
 
-JNI_EXPORT jint JNI_AppWebMessagePortDescriptor_TakeHandleToEntangle(
+JNI_EXPORT jlong JNI_AppWebMessagePortDescriptor_TakeHandleToEntangle(
     JNIEnv* env,
     jlong native_message_port_decriptor) {
   blink::MessagePortDescriptor* message_port_descriptor =
@@ -83,7 +83,7 @@ JNI_EXPORT jint JNI_AppWebMessagePortDescriptor_TakeHandleToEntangle(
 JNI_EXPORT void JNI_AppWebMessagePortDescriptor_GiveDisentangledHandle(
     JNIEnv* env,
     jlong native_message_port_decriptor,
-    jint native_handle) {
+    jlong native_handle) {
   blink::MessagePortDescriptor* message_port_descriptor =
       reinterpret_cast<blink::MessagePortDescriptor*>(
           native_message_port_decriptor);
