+++ mojo/public/java/system/javatests/src/org/chromium/mojo/HandleMock.java
@@ -78,7 +78,7 @@ public class HandleMock implements UntypedHandle, MessagePipeHandle, ProducerHan
      * @see Handle#releaseNativeHandle()
      */
     @Override
-    public int releaseNativeHandle() {
+    public long releaseNativeHandle() {
         return 0;
     }
