+++ mojo/public/java/system/src/org/chromium/mojo/system/impl/HandleBase.java
@@ -20,7 +20,7 @@ abstract class HandleBase implements Handle {
     /**
      * The pointer to the scoped handle owned by this object.
      */
-    private int mMojoHandle;
+    private long mMojoHandle;
 
     /**
      * The core implementation. Will be used to delegate all behavior.
@@ -30,7 +30,7 @@ abstract class HandleBase implements Handle {
     /**
      * Base constructor. Takes ownership of the passed handle.
      */
-    HandleBase(CoreImpl core, int mojoHandle) {
+    HandleBase(CoreImpl core, long mojoHandle) {
         mCore = core;
         mMojoHandle = mojoHandle;
     }
@@ -42,7 +42,7 @@ abstract class HandleBase implements Handle {
     protected HandleBase(HandleBase other) {
         mCore = other.mCore;
         HandleBase otherAsHandleImpl = other;
-        int mojoHandle = otherAsHandleImpl.mMojoHandle;
+        long mojoHandle = otherAsHandleImpl.mMojoHandle;
         otherAsHandleImpl.mMojoHandle = CoreImpl.INVALID_HANDLE;
         mMojoHandle = mojoHandle;
     }
@@ -54,7 +54,7 @@ abstract class HandleBase implements Handle {
     public void close() {
         if (mMojoHandle != CoreImpl.INVALID_HANDLE) {
             // After a close, the handle is invalid whether the close succeed or not.
-            int handle = mMojoHandle;
+            long handle = mMojoHandle;
             mMojoHandle = CoreImpl.INVALID_HANDLE;
             mCore.close(handle);
         }
@@ -96,8 +96,8 @@ abstract class HandleBase implements Handle {
      * @see Handle#releaseNativeHandle()
      */
     @Override
-    public int releaseNativeHandle() {
-        int result = mMojoHandle;
+    public long releaseNativeHandle() {
+        long result = mMojoHandle;
         mMojoHandle = CoreImpl.INVALID_HANDLE;
         return result;
     }
@@ -107,7 +107,7 @@ abstract class HandleBase implements Handle {
      *
      * @return the native scoped handle.
      */
-    int getMojoHandle() {
+    long getMojoHandle() {
         return mMojoHandle;
     }
